refactor(coding-agent): extract gh run module
Move the run_watch domain (Actions run/job snapshots, poll loop, watch/commit formatters and detail builders, log fetching) into gh-run; facade dispatch imports executeRunWatch.

Op: compress
